Remove MappingCollector-Delegations to super class
